Reorganise react to modules
