+++ web/reactjs/src/modules/Auth/container/Auth.js
@@ -1,0 +1,37 @@
+// Copyright 2020 Paul Sitoh
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import { connect } from 'react-redux';
+import { Auth } from '../components';
+import { fetchAuthUser } from '../../actions';
+
+const mapStateToProps = (store) => {
+    return {
+        user: store.authReducer.user
+    };
+};
+
+const mapDispatchToProps = (dispatch) => {
+    return {
+        authenticate: async (userName, password) => dispatch({
+            type: 'AUTH_USER_FULFILLED', 
+            payload: await fetchAuthUser(userName, password),
+        })
+    };
+};
+
+export default connect(
+    mapStateToProps,
+    mapDispatchToProps
+)(Auth);
