+++ web/reactjs/src/modules/MainLayout/components/Sidebar/Profile/Profile.js
