+++ web/reactjs/src/modules/MainLayout/components/Sidebar/Profile/index.js
