+++ web/reactjs/src/modules/MainLayout/components/Sidebar/Sidebar.js
