+++ web/reactjs/src/modules/MinimalLayout/components/MinimalLayout.js
