+++ web/reactjs/src/modules/MinimalLayout/components/Topbar/Topbar.js
