+++ web/reactjs/src/modules/MinimalLayout/components/index.js
