+++ web/reactjs/src/modules/MinimalLayout/index.js
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export { default } from './Auth';
+export { default } from './components';
