+++ web/reactjs/src/modules/NotFound/index.js
@@ -12,5 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export { default as Auth } from './Auth';
-export { default as MainLayout } from './MainLayout';
+export { default } from './components';
